Channel events
When registering server-side handlers for events targetting a channel (PublishingEvent, SubscribedEvent, SubscribingEvent, UnsubscribedEvent), it is now possible to specify a channel, to which is the handler related. Wildcards (*, **) are supported.

The chat demo has been extended with a whisper function, to demonstrate the usage.

CometdTerminal.html has been added, as a tool for testing the comet server.